Remove try block, Move tests around
