+++ data_ingest/tests/test_ingestors.py
@@ -46,9 +46,10 @@ class TestRowwiseValidator(SimpleTestCase):
 
 class TestValidationOutput(SimpleTestCase):
     def test_add_row_error(self):
-        output = ValidatorOutput(headers=["column1", "column2", "column3"],
+        output = ValidatorOutput(
                                  [(1, {"column1": "abc", "column2": "efg", "column3": "hij"}),
-                                  (2, {"column1": "xyz", "column2": "klm", "column3": "nop"})])
+                                  (2, {"column1": "xyz", "column2": "klm", "column3": "nop"})],
+                                  headers=["column1", "column2", "column3"])
 
         output.add_row_error(1, "Error", "E12", "Incorrect data", ["column1", "column3"])
         self.assertEqual(len(output.row_errors.items()), 1)
@@ -70,9 +71,10 @@ class TestValidationOutput(SimpleTestCase):
                              {"severity": "Info", "code": "I30", "message": "Information", "error_columns": []})
 
     def test_create_rows(self):
-        output = ValidatorOutput(headers=["column1", "column2", "column3"],
+        output = ValidatorOutput(
                                  OrderedDict([(1, {"column1": "abc", "column2": "efg", "column3": "hij"}),
-                                              (2, {"column1": "xyz", "column2": "klm", "column3": "nop"})]))
+                                              (2, {"column1": "xyz", "column2": "klm", "column3": "nop"})]),
+                                 headers=["column1", "column2", "column3"])
 
         output.row_errors = {
             1: [{'severity': 'Error', 'code': 'E12', 'message': 'Incorrect data',
@@ -100,9 +102,10 @@ class TestValidationOutput(SimpleTestCase):
         self.assertDictEqual(rows[1]["data"], {"column1": "xyz", "column2": "klm", "column3": "nop"})
 
     def test_get_output(self):
-        output = ValidatorOutput(headers=["column1", "column2", "column3"],
+        output = ValidatorOutput(
                                  OrderedDict([(1, {"column1": "abc", "column2": "efg", "column3": "hij"}),
-                                              (2, {"column1": "xyz", "column2": "klm", "column3": "nop"})]))
+                                              (2, {"column1": "xyz", "column2": "klm", "column3": "nop"})]),
+                                 headers=["column1", "column2", "column3"],)
 
         output.row_errors = {
             1: [{'severity': 'Error', 'code': 'E12', 'message': 'Incorrect data',
